+++ api/api-private-cluster.go
@@ -153,7 +153,7 @@ func (ps *privateServer) ClusterRouterRefresh(ctx context.Context, in *pb.AdminE
 }
 
 // ClusterStorageGroupUsage returns Storage Group Tenant's bucket usage report in a defined period of time
-func (ps *privateServer) ClusterStorageGroupUsage(ctx context.Context, in *pb.StorageGroupUsageRequest) (*pb.StorageGroupUsageResponse, error) {
+func (ps *privateServer) ClusterStorageGroupUsage(ctx context.Context, in *pb.StorageGroupReportRequest) (*pb.StorageGroupUsageResponse, error) {
 	scName := in.GetStorageCluster()
 	sgName := in.GetStorageGroup()
 	fromDate := in.GetFromDate()
@@ -219,4 +219,76 @@ func (ps *privateServer) ClusterStorageGroupUsage(ctx context.Context, in *pb.St
 		}
 	}
 	return &pb.StorageGroupUsageResponse{Usage: metrics}, nil
+}
+
+// ClusterStorageGroupSummary returns Storage Group Tenant's summary in a defined period of time
+// 	It includes, total number of users, total number of service accounts, total number of buckets and Average Usage per bucket
+func (ps *privateServer) ClusterStorageGroupSummary(ctx context.Context, in *pb.StorageGroupReportRequest) (*pb.StorageGroupSummaryResponse, error) {
+	scName := in.GetStorageCluster()
+	sgName := in.GetStorageGroup()
+	fromDate := in.GetFromDate()
+	toDate := in.GetToDate()
+
+	// Validate dates
+	layout := cluster.PostgresShortTimeLayout
+	fromDateFormatted, err := time.Parse(layout, fromDate)
+	if err != nil {
+		log.Println("Wrong date format:", err)
+		return nil, status.New(codes.InvalidArgument, "wrong date format").Err()
+	}
+	toDateFormatted, err := time.Parse(layout, toDate)
+	if err != nil {
+		log.Println("Wrong date format:", err)
+		return nil, status.New(codes.InvalidArgument, "wrong date format").Err()
+	}
+
+	appCtx, err := cluster.NewEmptyContextWithGrpcContext(ctx)
+	if err != nil {
+		log.Println(err)
+		return nil, status.New(codes.Internal, "Internal error").Err()
+	}
+
+	// fetch storage cluster and storage group
+	storageCluster, err := cluster.GetStorageClusterByName(appCtx, scName)
+	if err != nil {
+		log.Println("Error getting storage cluster by name:", err)
+		if err == sql.ErrNoRows {
+			return nil, status.New(codes.NotFound, "Storage Cluster not found").Err()
+		}
+		return nil, status.New(codes.Internal, "Internal error").Err()
+	}
+	storageGroup, err := cluster.GetStorageGroupByNameNStorageCluster(appCtx, sgName, storageCluster)
+	if err != nil {
+		log.Println("Error getting storage group by name:", err)
+		if err == sql.ErrNoRows {
+			return nil, status.New(codes.NotFound, "Storage Group not found").Err()
+		}
+		return nil, status.New(codes.Internal, "Internal error").Err()
+	}
+
+	// Get all services(tenants) of a storage group and only get the tenant bucket usage if the service has been claimed
+	tenants := <-cluster.GetListOfTenantsForStorageGroup(appCtx, storageGroup)
+	var metrics []*pb.StorageGroupSummary
+	for _, tenant := range tenants {
+		if !tenant.Available {
+			tenantDB := db.GetInstance().GetTenantDB(tenant.Tenant.ShortName)
+			bucketUsageMetrics, err := cluster.GetTenantsSummaryDb(tenantDB, fromDateFormatted, toDateFormatted)
+			if err != nil {
+				log.Println("error getting daily bucket metrics:", err)
+				return nil, status.New(codes.Internal, "Internal error").Err()
+			}
+			for _, sm := range bucketUsageMetrics {
+				metric := &pb.StorageGroupSummary{
+					Date:                 sm.Time.String(),
+					Tenant:               tenant.Tenant.Name,
+					TotalUsers:           sm.UsersCount,
+					TotalServiceAccounts: sm.ServiceAccountsCount,
+					TotalBuckets:         sm.BucketsCount,
+					Usage:                sm.Usage,
+				}
+				metrics = append(metrics, metric)
+			}
+		}
+	}
+	return &pb.StorageGroupSummaryResponse{Summary: metrics}, nil
 }