+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Enabled TLS communication (`gRPC` and `HTTPS`) between `nginx-resolver`, `portal-proxy`, `m3-portal-frontend`, `m3-portal-backend` and `m3` services using [autocert](https://github.com/smallstep/autocert)
-- Cli command to get storage group tenant's bucket's average usage in a defined period `./m3 cluster sc sg usage`.
+- Cli command to get storage group tenant's summary in a defined period `./m3 cluster sc sg summary`.
 - Start using "changelog" file.
 
 ### Changed